Prevent output of empty cells
The csv.reader puts in empty strings for empty cells.  At the
point of generating the filepath, change these entries to None.

Add exception for empty keyword entries or column numbers out of
range during filepath generation.

Currently, an empty cell referenced by index will still replace
with None in the filepath.  This should be amended in the future
with a subclass of format.