Aspect ratios with colons
The function that santizes file names was clobbering aspect
ratios that used a colon.  So if a csv entry included 1:1, it
would be come 1_1 in the destination file name.

Now that function will scan for aspect rations using colons first.
If found will ch`ange from : to an x (1:1 to 1x1).  Then proceed on
with santizing the remaining symbols, duplicates, etc.